chore: refactor flutter_acrylic.dart
